+++ src/app/components/file-chunks-bar.js
@@ -8,19 +8,43 @@ const {
   computed,
 } = Ember;
 
+const barWidth = 320;
+const barHeight = 20;
+
 /**
  * Draw a bar with file chunks (blocks) using function from utils.
  * @module components/file-chunks-bar
  * @author Jakub Liput
- * @copyright (C) 2016 ACK CYFRONET AGH
+ * @copyright (C) 2016-2018 ACK CYFRONET AGH
  * @license This software is released under the MIT license cited in 'LICENSE.txt'.
  */
 export default Ember.Component.extend({
   classNames: ['file-chunks-bar'],
   
+  /**
+   * @virtual
+   * Needs:
+   * - id
+   * - sizeHumanReadable
+   * - size
+   */
   file: null,
-  fileBlocks: null,
-
+  
+  /**
+   * @virtual
+   * Format: `{ start px of bar [0-319]: green opacity [0-100] }`
+   * @type {object}
+   */
+  chunksBarData: undefined,
+  
+  /**
+   * @virtual
+   * Percentage of blocks of the file on this provider.
+   * Float in range 0..100.
+   * @type {Number}
+   */
+  blocksPercentage: undefined,
+  
   didInsertElement() {
     this._super(...arguments);
     run.scheduleOnce('afterRender', this, function() {
@@ -34,42 +58,38 @@ export default Ember.Component.extend({
   isLoading: false,
 
   /**
-   * @type {Ember.ComputedProperty<number>}
+   * @type {Ember.ComputedProperty<string>}
    */
-  chunksPercent: computed('file.size', 'fileBlocks.blocks', function () {
-    const size = this.get('file.size');
-    const blocks = this.get('fileBlocks.blocks');
-    if (size && blocks) {
-      let sum = 0;
-      for (let i = 0; i < blocks.length; i += 2) {
-        sum += blocks[i + 1] - blocks[i];
-      }
-      return Math.floor(sum / size * 1000) / 10;
+  chunksPercent: computed('blocksPercentage', function chunksPercent() {
+    const blocksPercentage = this.get('blocksPercentage') || 0;
+    if (blocksPercentage > 0 && blocksPercentage < 0.1) {
+      return '< 0.1%';
     } else {
-      return undefined;
+      return `${Math.round(blocksPercentage * 10) / 10}%`;
     }
   }),
 
   // should use new everytime?
-  redrawCanvas: observer('$canvas', 'file.size', 'fileBlocks.blocks', function() {
+  redrawCanvas: observer('$canvas', 'file.size', 'chunksBarData', function() {
     let {
       file,
-      fileBlocks,
+      chunksBarData,
       $canvas
-    } = this.getProperties('$canvas', 'file', 'fileBlocks', 'canvas');
+    } = this.getProperties('$canvas', 'file', 'chunksBarData', 'canvas');
 
     let size = get(file, 'size');
-    let blocks = get(fileBlocks, 'blocks');
 
-    if (size && blocks) {
+    if (size && chunksBarData) {
       try {
         this.setProperties({
           isRenderFailed: false,
           isLoading: true
         });
         new FileChunksBar($canvas, {
-          file_size: size,
-          chunks: blocks
+          fileSize: size,
+          chunksBarData,
+          width: barWidth,
+          height: barHeight,
         });
         this.setProperties({
           isRendered: true,