+++ src/tests/unit/models/file-distribution-test.js
@@ -12,7 +12,7 @@ describeModel(
   function() {
     it('computes if there are no file blocks', function () {
       const model = this.subject({
-        blocks: [],
+        blocksPercentage: 0,
         fileSize: FILE_SIZE,
       });
       expect(model.get('isEmpty')).to.be.true;
@@ -20,7 +20,7 @@ describeModel(
     
     it('computes if file distribution is not empty nor complete', function() {
       const model = this.subject({
-        blocks: [10, 20],
+        blocksPercentage: 30,
         fileSize: FILE_SIZE,
       });
       expect(model.get('isEmpty')).to.be.false;
@@ -29,26 +29,10 @@ describeModel(
     
     it('computes if file distribution is complete', function () {
       const model = this.subject({
-        blocks: [0, 10, 10, 90, 90, FILE_SIZE],
+        blocksPercentage: 100,
         fileSize: FILE_SIZE,
       });
       expect(model.get('isComplete')).to.be.true;
-    });
-    
-    it('detects that file is empty even if it has entries in block array', function () {
-      const model = this.subject({
-        blocks: [0, 0, 10, 10],
-        fileSize: FILE_SIZE,
-      });
-      expect(model.get('isEmpty')).to.be.true;
-    });
-    
-    it('detects that file is not complete if it has "empty" entries', function () {
-      const model = this.subject({
-        blocks: [0, 0],
-        fileSize: FILE_SIZE,
-      });
-      expect(model.get('isComplete')).to.be.false;
     });
   }
 );